# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dedicated API for ArcheryHub Mobile App (Archer \u0026 Scorekeeper)",
        "title": "ArcheryHub Mobile API",
        "termsOfService": "http://archeryhub.id/terms/",
        "contact": {
            "name": "ArcheryHub Support",
            "url": "https://archeryhub.id",
            "email": "support@archeryhub.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1"
    },
    "host": "api.archeryhub.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/mobile/marketplace/products": {
            "get": {
                "description": "Get a list of active products in the marketplace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Marketplace"
                ],
                "summary": "List Marketplace Products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileMarketplaceProductsResponse"
                        }
                    }
                }
            }
        },
        "/mobile/marketplace/products/{id}": {
            "get": {
                "description": "Get detailed information about a specific product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Marketplace"
                ],
                "summary": "Get Product Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileMarketplaceProductResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/news": {
            "get": {
                "description": "Get a list of published news articles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - News"
                ],
                "summary": "List News",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by title or content",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileNewsListResponse"
                        }
                    }
                }
            }
        },
        "/mobile/news/{id}": {
            "get": {
                "description": "Get the full content of a news article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - News"
                ],
                "summary": "Get News Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileNewsDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/news/{id}/comments": {
            "get": {
                "description": "Get approved comments for a news article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - News"
                ],
                "summary": "List News Comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileNewsCommentsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Post a new comment as an authenticated user or guest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - News"
                ],
                "summary": "Add News Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileAddNewsCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mobile.MobileAddNewsCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Wah acaranya seru banget!"
                },
                "guest_name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "user_id": {
                    "type": "string",
                    "example": "arc-a49ee7d7-9d7b-4be7-8652-342f2fca23f9"
                }
            }
        },
...
        "mobile.MobileMarketplaceProductResponse": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/models.Product"
                }
            }
        },
        "mobile.MobileMarketplaceProductsResponse": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "mobile.MobileNewsComment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "guest_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "news_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "mobile.MobileNewsCommentsResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mobile.MobileNewsComment"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "mobile.MobileNewsDetail": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "meta_description": {
                    "type": "string"
                },
                "meta_title": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "mobile.MobileNewsDetailResponse": {
            "type": "object",
            "properties": {
                "news": {
                    "$ref": "#/definitions/mobile.MobileNewsDetail"
                }
            }
        },
        "mobile.MobileNewsItem": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "mobile.MobileNewsListResponse": {
            "type": "object",
            "properties": {
                "news": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mobile.MobileNewsItem"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "equipment, apparel, accessories, training, other",
                    "type": "string"
                },
                "club_id": {
                    "type": "string"
                },
                "colors": {
                    "description": "JSON array",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "images": {
                    "description": "JSON array",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "sale_price": {
                    "type": "number"
                },
                "seller_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "specifications": {
                    "description": "JSON object",
                    "type": "string"
                },
                "status": {
                    "description": "draft, active, sold_out, archived",
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}